+++ CHANGES.md
@@ -34,6 +34,10 @@
 * Interfaces may now be parameterized by other interfaces (interface functors).
   ([#1582](https://github.com/GaloisInc/cryptol/issues/1582))
 
+* Add module aliases for giving short names to
+  existing modules, functors, or interfaces.
+  ([#1591](https://github.com/GaloisInc/cryptol/issues/1591))
+
 ## Bug fixes
 
 * Fix pretty printing of types in errors messages